Fix split package with voting-only nodes
Moves the implementation of voting-only nodes to the
`o.e.c.c.votingonly` package.
