+++ x-pack/plugin/voting-only-node/src/test/java/org/elasticsearch/cluster/coordination/votingonly/LocalStateVotingOnlyNodePlugin.java
@@ -4,7 +4,7 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-package org.elasticsearch.cluster.coordination;
+package org.elasticsearch.cluster.coordination.votingonly;
 
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.xpack.core.LocalStateCompositeXPackPlugin;
